Teste de chat finalizado!

+++ chatbot.py
@@ -3,7 +3,6 @@ from dotenv import load_dotenv
 from langchain.schema import (HumanMessage)
 from langchain_groq import ChatGroq
 from pinecone import Pinecone, ServerlessSpec
-import re
 from sentence_transformers import SentenceTransformer
 from datetime import datetime
 
@@ -13,14 +12,15 @@ class ChatBot:
     def __init__(self):
         self.groq_api_key, self.openai_api_key, self.db_api_key = self.get_env_info()
         self.chat = ChatGroq(temperature=0, model_name="llama-3.1-70b-versatile")
-        self.chat_classifier_continuous_learning = self.chat
-        self.chat_classifier_valid_correction = self.chat
+        self.chat_classifier_continuous_learning = self.chat_classifier_valid_correction = self.chat
         self.db = Pinecone(api_key=self.db_api_key, environment="us-west1-gcp")
         self.embeddings_model = SentenceTransformer('bert-base-nli-mean-tokens')
         self.messages = []
         self.checking_indexes()
         self.index_preferences = self.get_index('index-preferences')
         self.index_corrections = self.get_index('index-corrections')
+        self.prev_question = self.prev_answer = None
+        self.restarting_indexes()
 
     @staticmethod
     def get_env_info():
@@ -33,6 +33,10 @@ class ChatBot:
     def checking_indexes(self):
         self.creating_index_if_it_does_not_exists('index-preferences')
         self.creating_index_if_it_does_not_exists('index-corrections')
+
+    def restarting_indexes(self):
+        self.erase_index_content(self.index_preferences)
+        self.erase_index_content(self.index_corrections)
 
     def creating_index_if_it_does_not_exists(self, index_name):
         existing_indexes = [i['name'] for i in self.db.list_indexes().to_dict()['indexes']]
@@ -75,12 +79,12 @@ class ChatBot:
 
         return preference_class
 
-    def classifier_valid_correction(self, user_correction, prev_question, prev_answer):
+    def classifier_valid_correction(self, user_correction):
         valid_correction = False
-        if prev_answer is not None and prev_question is not None:
+        if self.prev_answer is not None and self.prev_question is not None:
             prompt = (f"Considerando o diálogo abaixo, a correção do usuário está correta? Responda com Sim ou Não: \n"
-                      f"Usuário: {prev_question} \n"
-                      f"Resposta: {prev_answer}' \n"
+                      f"Usuário: {self.prev_question} \n"
+                      f"Resposta: {self.prev_answer}' \n"
                       f"Usuário: {user_correction}")
             corrections_labels = {'sim': True, 'não': False}
             detected_classes = self.chat_classifier_valid_correction.invoke(prompt)
@@ -88,19 +92,18 @@ class ChatBot:
             valid_correction = corrections_labels.get(answer, False)
         return valid_correction
 
-    def interaction(self, user_prompt, prev_question, prev_answer):
-        self.write_prompt(user_prompt, prev_question, prev_answer)
+    def interaction(self, user_prompt):
+        self.write_prompt(user_prompt)
         ai_answer = self.informed_chat_answer(user_prompt)
         print(ai_answer)
         return ai_answer
 
-    def write_prompt(self, user_prompt, prev_question, prev_answer):
+    def write_prompt(self, user_prompt):
+        self.messages.append(user_prompt)
         is_preference = self.classifier_preference(HumanMessage(user_prompt))
         if is_preference:
-            print('Is preference')
             self.store_interaction_in_db(user_prompt, self.index_preferences, 'preference')
-        if self.classifier_valid_correction(user_prompt, prev_question, prev_answer):
-            print('Is correction')
+        if self.classifier_valid_correction(user_prompt):
             self.store_interaction_in_db(user_prompt, self.index_corrections, 'correction')
 
     def store_interaction_in_db(self, user_prompt, index, interaction):
@@ -144,24 +147,28 @@ class ChatBot:
                 if relevant_responses_corrections else context
 
         prompt = f"Baseado no seguinte contexto:\n{context}\n\nPergunta: {user_query}" if context != '' else user_query
-        return prompt
+        self.messages.append(prompt)
+        self.messages = list(set(self.messages))
 
     def informed_chat_answer(self, user_message):
-        prompt = self.prompt_for_retrieving_relevant_info(user_message)
-        ai_answer = self.chat.invoke(prompt)
-        return ai_answer.content
+        self.prompt_for_retrieving_relevant_info(user_message)
+        ai_answer = self.chat.invoke(self.messages).content
+        self.messages.append(ai_answer)
+        return ai_answer
 
     def main(self):
-        prev_question = prev_answer = None
         while True:
             user_prompt = input("Talk to me!")
-            ai_answer = self.interaction(user_prompt, prev_question, prev_answer)
-            prev_question = user_prompt
-            prev_answer = ai_answer
+            ai_answer = self.interaction(user_prompt)
+            self.prev_question = user_prompt
+            self.prev_answer = ai_answer
 
     @staticmethod
     def erase_index_content(index):
-        index.delete(delete_all=True)
+        stats_index = index.describe_index_stats().get("namespaces", {})
+        namespaces = stats_index.get("namespaces", {})
+        n_vectors = sum(ns.get("vector_count", 0) for ns in namespaces.values())
+        index.delete(delete_all=True) if n_vectors != 0 else None
 
 
 a = ChatBot()
